+++ R/kNN.plot.R
@@ -12,9 +12,12 @@
 #     k-Nearest Neighbour Classification using formula interface
 ## - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - |
 
-kNN = function( formula, train, test, k = 1, transform = FALSE, l = 0, 
-                prob = FALSE, use.all = TRUE, na.rm = FALSE ) 
+kNN.plot = function( formula, train, test, k.max = 10, transform = FALSE, set.seed = NULL, ... ) 
 {
+    if( k.max < 2 ) stop( "  k.max must be > 1" )
+    
+    if( !is.null( set.seed ) ) set.seed( set.seed )
+    
     formula = stats::as.formula( formula )
     
     model_frame_train  = stats::model.frame( formula, data = train )
@@ -22,6 +25,7 @@ kNN = function( formula, train, test, k = 1, transform = FALSE, l = 0,
     model_train_no_res = stats::delete.response( model_train )
     
     train_label = stats::model.response( model_frame_train )
+    test_label  = stats::model.matrix( model_train_no_res, test )[ , 1 ]
     
     train = stats::model.matrix( model_train, model_frame_train )[ , -1 ]
     test  = stats::model.matrix( model_train_no_res, test )[ , -1 ]
@@ -45,10 +49,31 @@ kNN = function( formula, train, test, k = 1, transform = FALSE, l = 0,
         }
     }
     
-    class::knn( train = train, test = test, cl = train_label, k = k, l = l, 
-                prob = prob, use.all = use.all )
+    k_list   = 1:k.max
+    mse_list = vector( length = k.max  )
+    
+    for( k in k_list ){
+        knn_k = class::knn( train = train, test = test, cl = train_label, k = k, ... )
+        
+        mse_list[ k ] = liver::mse( knn_k, test_label )
+    }
+    
+    df = data.frame( k_list = as.factor( k_list ), mse_list = mse_list, stringsAsFactors = TRUE)
+    
+    df_gg = data.frame( k_list = k_list, mse_list = mse_list )
+    
+    ggplot2::ggplot( df_gg, ggplot2::aes( x = k_list, y = mse_list ) ) +
+        ggplot2::geom_line( color = "#a0a0a0" ) + 
+        ggplot2::geom_point( shape = 21, color = "#ff5085", fill = "#ff83a8", size = 2 ) + 
+        ggplot2::theme_minimal() + 
+        ggplot2::scale_x_continuous( breaks = k_list ) +
+        ggplot2::ggtitle( "Optimal value of k based on MSE" ) +
+        ggplot2::labs( x = "Value of k", y = "Mean Square Error (MSE)" ) +
+        ggplot2::theme( axis.line = ggplot2::element_line( size = 0.4, colour = "black" ),
+               panel.grid.major = ggplot2::element_line( colour = "#f2f2f2" ), panel.grid.minor = ggplot2::element_blank(),
+               panel.border = ggplot2::element_blank(), panel.background = ggplot2::element_blank() )
 }
-
+   
 ## - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - |